04-data-storage/requirements.txt
# CosmosDB Dependencies
azure-cosmos==4.5.1
python-dotenv==1.0.0

